csharpier example

+++ bitwarden_license/src/Commercial.Infrastructure.EntityFramework/SecretsManager/Repositories/ServiceAccountRepository.cs
@@ -10,13 +10,18 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Bit.Commercial.Infrastructure.EntityFramework.SecretsManager.Repositories;
 
-public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.ServiceAccount, ServiceAccount, Guid>, IServiceAccountRepository
+public class ServiceAccountRepository
+    : Repository<Core.SecretsManager.Entities.ServiceAccount, ServiceAccount, Guid>,
+        IServiceAccountRepository
 {
     public ServiceAccountRepository(IServiceScopeFactory serviceScopeFactory, IMapper mapper)
-        : base(serviceScopeFactory, mapper, db => db.ServiceAccount)
-    { }
+        : base(serviceScopeFactory, mapper, db => db.ServiceAccount) { }
 
-    public async Task<IEnumerable<Core.SecretsManager.Entities.ServiceAccount>> GetManyByOrganizationIdAsync(Guid organizationId, Guid userId, AccessClientType accessType)
+    public async Task<IEnumerable<Core.SecretsManager.Entities.ServiceAccount>> GetManyByOrganizationIdAsync(
+        Guid organizationId,
+        Guid userId,
+        AccessClientType accessType
+    )
     {
         using var scope = ServiceScopeFactory.CreateScope();
         var dbContext = GetDatabaseContext(scope);
@@ -37,13 +42,13 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
     {
         using var scope = ServiceScopeFactory.CreateScope();
         var dbContext = GetDatabaseContext(scope);
-        var serviceAccounts = await dbContext.ServiceAccount
-            .Where(c => ids.Contains(c.Id))
-            .ToListAsync();
+        var serviceAccounts = await dbContext.ServiceAccount.Where(c => ids.Contains(c.Id)).ToListAsync();
         return Mapper.Map<List<Core.SecretsManager.Entities.ServiceAccount>>(serviceAccounts);
     }
 
-    public async Task<IEnumerable<Core.SecretsManager.Entities.ServiceAccount>> GetManyByOrganizationIdWriteAccessAsync(Guid organizationId, Guid userId, AccessClientType accessType)
+    public async Task<
+        IEnumerable<Core.SecretsManager.Entities.ServiceAccount>
+    > GetManyByOrganizationIdWriteAccessAsync(Guid organizationId, Guid userId, AccessClientType accessType)
     {
         using var scope = ServiceScopeFactory.CreateScope();
         var dbContext = GetDatabaseContext(scope);
@@ -69,26 +74,27 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         await using var transaction = await dbContext.Database.BeginTransactionAsync();
 
         // Policies can't have a cascade delete, so we need to delete them manually.
-        await dbContext.AccessPolicies.Where(ap =>
-                targetIds.Contains(((ServiceAccountProjectAccessPolicy)ap).ServiceAccountId!.Value) ||
-                targetIds.Contains(((ServiceAccountSecretAccessPolicy)ap).ServiceAccountId!.Value) ||
-                targetIds.Contains(((GroupServiceAccountAccessPolicy)ap).GrantedServiceAccountId!.Value) ||
-                targetIds.Contains(((UserServiceAccountAccessPolicy)ap).GrantedServiceAccountId!.Value))
+        await dbContext
+            .AccessPolicies.Where(ap =>
+                targetIds.Contains(((ServiceAccountProjectAccessPolicy)ap).ServiceAccountId!.Value)
+                || targetIds.Contains(((ServiceAccountSecretAccessPolicy)ap).ServiceAccountId!.Value)
+                || targetIds.Contains(((GroupServiceAccountAccessPolicy)ap).GrantedServiceAccountId!.Value)
+                || targetIds.Contains(((UserServiceAccountAccessPolicy)ap).GrantedServiceAccountId!.Value)
+            )
             .ExecuteDeleteAsync();
 
-        await dbContext.ApiKeys
-            .Where(a => targetIds.Contains(a.ServiceAccountId!.Value))
-            .ExecuteDeleteAsync();
+        await dbContext.ApiKeys.Where(a => targetIds.Contains(a.ServiceAccountId!.Value)).ExecuteDeleteAsync();
 
-        await dbContext.ServiceAccount
-            .Where(c => targetIds.Contains(c.Id))
-            .ExecuteDeleteAsync();
+        await dbContext.ServiceAccount.Where(c => targetIds.Contains(c.Id)).ExecuteDeleteAsync();
 
         await transaction.CommitAsync();
     }
 
-    public async Task<(bool Read, bool Write)> AccessToServiceAccountAsync(Guid id, Guid userId,
-        AccessClientType accessType)
+    public async Task<(bool Read, bool Write)> AccessToServiceAccountAsync(
+        Guid id,
+        Guid userId,
+        AccessClientType accessType
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
@@ -104,7 +110,8 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
     public async Task<Dictionary<Guid, (bool Read, bool Write)>> AccessToServiceAccountsAsync(
         IEnumerable<Guid> ids,
         Guid userId,
-        AccessClientType accessType)
+        AccessClientType accessType
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
@@ -120,13 +127,15 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         using (var scope = ServiceScopeFactory.CreateScope())
         {
             var dbContext = GetDatabaseContext(scope);
-            return await dbContext.ServiceAccount
-                .CountAsync(ou => ou.OrganizationId == organizationId);
+            return await dbContext.ServiceAccount.CountAsync(ou => ou.OrganizationId == organizationId);
         }
     }
 
-    public async Task<int> GetServiceAccountCountByOrganizationIdAsync(Guid organizationId, Guid userId,
-        AccessClientType accessType)
+    public async Task<int> GetServiceAccountCountByOrganizationIdAsync(
+        Guid organizationId,
+        Guid userId,
+        AccessClientType accessType
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
@@ -142,8 +151,11 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         return await query.CountAsync();
     }
 
-    public async Task<ServiceAccountCounts> GetServiceAccountCountsByIdAsync(Guid serviceAccountId, Guid userId,
-        AccessClientType accessType)
+    public async Task<ServiceAccountCounts> GetServiceAccountCountsByIdAsync(
+        Guid serviceAccountId,
+        Guid userId,
+        AccessClientType accessType
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
@@ -160,24 +172,37 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         {
             Projects = serviceAccount.ProjectAccessPolicies.Count,
             People = serviceAccount.UserAccessPolicies.Count + serviceAccount.GroupAccessPolicies.Count,
-            AccessTokens = serviceAccount.ApiKeys.Count
+            AccessTokens = serviceAccount.ApiKeys.Count,
         });
 
         var serviceAccountCounts = await serviceAccountCountsQuery.FirstOrDefaultAsync();
-        return serviceAccountCounts ?? new ServiceAccountCounts { Projects = 0, People = 0, AccessTokens = 0 };
+        return serviceAccountCounts
+            ?? new ServiceAccountCounts
+            {
+                Projects = 0,
+                People = 0,
+                AccessTokens = 0,
+            };
     }
 
-    public async Task<bool> ServiceAccountsAreInOrganizationAsync(List<Guid> serviceAccountIds, Guid organizationId)
+    public async Task<bool> ServiceAccountsAreInOrganizationAsync(
+        List<Guid> serviceAccountIds,
+        Guid organizationId
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
         var result = await dbContext.ServiceAccount.CountAsync(sa =>
-            sa.OrganizationId == organizationId && serviceAccountIds.Contains(sa.Id));
+            sa.OrganizationId == organizationId && serviceAccountIds.Contains(sa.Id)
+        );
         return serviceAccountIds.Count == result;
     }
 
     public async Task<IEnumerable<ServiceAccountSecretsDetails>> GetManyByOrganizationIdWithSecretsDetailsAsync(
-        Guid organizationId, Guid userId, AccessClientType accessType)
+        Guid organizationId,
+        Guid userId,
+        AccessClientType accessType
+    )
     {
         await using var scope = ServiceScopeFactory.CreateAsyncScope();
         var dbContext = GetDatabaseContext(scope);
@@ -187,9 +212,11 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         {
             AccessClientType.NoAccessCheck => serviceAccountQuery,
             AccessClientType.User => serviceAccountQuery.Where(c =>
-                c.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read) ||
-                c.GroupAccessPolicies.Any(ap =>
-                    ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read))),
+                c.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read)
+                || c.GroupAccessPolicies.Any(ap =>
+                    ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read)
+                )
+            ),
             _ => throw new ArgumentOutOfRangeException(nameof(accessType), accessType, null),
         };
 
@@ -199,88 +226,113 @@ public class ServiceAccountRepository : Repository<Core.SecretsManager.Entities.
         var projectSecretsAccessResults = await projectSecretsAccessQuery.ToListAsync();
         var directSecretAccessResults = await directSecretAccessQuery.ToListAsync();
 
-        var applicableDirectSecretAccessResults = FilterDirectSecretAccessResults(projectSecretsAccessResults, directSecretAccessResults);
+        var applicableDirectSecretAccessResults = FilterDirectSecretAccessResults(
+            projectSecretsAccessResults,
+            directSecretAccessResults
+        );
 
-        var results = projectSecretsAccessResults.Concat(applicableDirectSecretAccessResults)
+        var results = projectSecretsAccessResults
+            .Concat(applicableDirectSecretAccessResults)
             .GroupBy(g => g.ServiceAccount)
-            .Select(g =>
-                new ServiceAccountSecretsDetails
-                {
-                    ServiceAccount = Mapper.Map<Core.SecretsManager.Entities.ServiceAccount>(g.Key),
-                    AccessToSecrets = g.Sum(x => x.SecretIds.Count())
-                }).OrderBy(c => c.ServiceAccount.RevisionDate).ToList();
+            .Select(g => new ServiceAccountSecretsDetails
+            {
+                ServiceAccount = Mapper.Map<Core.SecretsManager.Entities.ServiceAccount>(g.Key),
+                AccessToSecrets = g.Sum(x => x.SecretIds.Count()),
+            })
+            .OrderBy(c => c.ServiceAccount.RevisionDate)
+            .ToList();
 
         return results;
     }
 
     private record ServiceAccountAccess(Guid Id, bool Read, bool Write);
 
-    private static IQueryable<ServiceAccountAccess> BuildServiceAccountAccessQuery(IQueryable<ServiceAccount> serviceAccountQuery, Guid userId,
-        AccessClientType accessType) =>
+    private static IQueryable<ServiceAccountAccess> BuildServiceAccountAccessQuery(
+        IQueryable<ServiceAccount> serviceAccountQuery,
+        Guid userId,
+        AccessClientType accessType
+    ) =>
         accessType switch
         {
-            AccessClientType.NoAccessCheck => serviceAccountQuery.Select(sa => new ServiceAccountAccess(sa.Id, true, true)),
-            AccessClientType.User => serviceAccountQuery.Select(sa => new ServiceAccountAccess
-            (
+            AccessClientType.NoAccessCheck => serviceAccountQuery.Select(sa => new ServiceAccountAccess(
                 sa.Id,
-                sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read) ||
-                sa.GroupAccessPolicies.Any(ap =>
-                    ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read)),
-                sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Write) ||
-                sa.GroupAccessPolicies.Any(ap =>
-                    ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Write))
+                true,
+                true
             )),
-            AccessClientType.ServiceAccount => serviceAccountQuery.Select(sa => new ServiceAccountAccess(sa.Id, false, false)),
-            _ => serviceAccountQuery.Select(sa => new ServiceAccountAccess(sa.Id, false, false))
+            AccessClientType.User => serviceAccountQuery.Select(sa => new ServiceAccountAccess(
+                sa.Id,
+                sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read)
+                    || sa.GroupAccessPolicies.Any(ap =>
+                        ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read)
+                    ),
+                sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Write)
+                    || sa.GroupAccessPolicies.Any(ap =>
+                        ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Write)
+                    )
+            )),
+            AccessClientType.ServiceAccount => serviceAccountQuery.Select(sa => new ServiceAccountAccess(
+                sa.Id,
+                false,
+                false
+            )),
+            _ => serviceAccountQuery.Select(sa => new ServiceAccountAccess(sa.Id, false, false)),
         };
 
-    private static Expression<Func<ServiceAccount, bool>> UserHasReadAccessToServiceAccount(Guid userId) => sa =>
-        sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read) ||
-        sa.GroupAccessPolicies.Any(ap => ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read));
+    private static Expression<Func<ServiceAccount, bool>> UserHasReadAccessToServiceAccount(Guid userId) =>
+        sa =>
+            sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Read)
+            || sa.GroupAccessPolicies.Any(ap =>
+                ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Read)
+            );
 
-    private static Expression<Func<ServiceAccount, bool>> UserHasWriteAccessToServiceAccount(Guid userId) => sa =>
-        sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Write) ||
-        sa.GroupAccessPolicies.Any(ap => ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Write));
+    private static Expression<Func<ServiceAccount, bool>> UserHasWriteAccessToServiceAccount(Guid userId) =>
+        sa =>
+            sa.UserAccessPolicies.Any(ap => ap.OrganizationUser.User.Id == userId && ap.Write)
+            || sa.GroupAccessPolicies.Any(ap =>
+                ap.Group.GroupUsers.Any(gu => gu.OrganizationUser.User.Id == userId && ap.Write)
+            );
 
-    private static IQueryable<ServiceAccountSecretsAccess> BuildProjectSecretsAccessQuery(DatabaseContext dbContext,
-        IQueryable<ServiceAccount> serviceAccountQuery) =>
+    private static IQueryable<ServiceAccountSecretsAccess> BuildProjectSecretsAccessQuery(
+        DatabaseContext dbContext,
+        IQueryable<ServiceAccount> serviceAccountQuery
+    ) =>
         from sa in serviceAccountQuery
-        join ap in dbContext.ServiceAccountProjectAccessPolicy
-            on sa.Id equals ap.ServiceAccountId into grouping
+        join ap in dbContext.ServiceAccountProjectAccessPolicy on sa.Id equals ap.ServiceAccountId into grouping
         from ap in grouping.DefaultIfEmpty()
-        select new ServiceAccountSecretsAccess
-        (
-            sa, ap.GrantedProject.Secrets.Where(s => s.DeletedDate == null).Select(s => s.Id)
+        select new ServiceAccountSecretsAccess(
+            sa,
+            ap.GrantedProject.Secrets.Where(s => s.DeletedDate == null).Select(s => s.Id)
         );
 
     private static IQueryable<ServiceAccountSecretsAccess> BuildDirectSecretAccessQuery(
         DatabaseContext dbContext,
-        IQueryable<ServiceAccount> serviceAccountQuery) =>
+        IQueryable<ServiceAccount> serviceAccountQuery
+    ) =>
         from sa in serviceAccountQuery
-        join ap in dbContext.ServiceAccountSecretAccessPolicy
-            on sa.Id equals ap.ServiceAccountId into grouping
+        join ap in dbContext.ServiceAccountSecretAccessPolicy on sa.Id equals ap.ServiceAccountId into grouping
         from ap in grouping.DefaultIfEmpty()
-        where ap.GrantedSecret.DeletedDate == null &&
-              ap.GrantedSecretId != null
-        select new ServiceAccountSecretsAccess(sa,
-            new List<Guid> { ap.GrantedSecretId!.Value });
+        where ap.GrantedSecret.DeletedDate == null && ap.GrantedSecretId != null
+        select new ServiceAccountSecretsAccess(sa, new List<Guid> { ap.GrantedSecretId!.Value });
 
     private static List<ServiceAccountSecretsAccess> FilterDirectSecretAccessResults(
         List<ServiceAccountSecretsAccess> projectSecretsAccessResults,
-        List<ServiceAccountSecretsAccess> directSecretAccessResults) =>
-        directSecretAccessResults.Where(directSecretAccessResult =>
-        {
-            var serviceAccountId = directSecretAccessResult.ServiceAccount.Id;
-            var secretId = directSecretAccessResult.SecretIds.FirstOrDefault();
-            if (secretId == Guid.Empty)
+        List<ServiceAccountSecretsAccess> directSecretAccessResults
+    ) =>
+        directSecretAccessResults
+            .Where(directSecretAccessResult =>
             {
-                return false;
-            }
+                var serviceAccountId = directSecretAccessResult.ServiceAccount.Id;
+                var secretId = directSecretAccessResult.SecretIds.FirstOrDefault();
+                if (secretId == Guid.Empty)
+                {
+                    return false;
+                }
 
-            return !projectSecretsAccessResults
-                .Where(x => x.ServiceAccount.Id == serviceAccountId)
-                .Any(x => x.SecretIds.Contains(secretId));
-        }).ToList();
+                return !projectSecretsAccessResults
+                    .Where(x => x.ServiceAccount.Id == serviceAccountId)
+                    .Any(x => x.SecretIds.Contains(secretId));
+            })
+            .ToList();
 
     private record ServiceAccountSecretsAccess(ServiceAccount ServiceAccount, IEnumerable<Guid> SecretIds);
 }
